# services/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manga Organizer サイドカー",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/pages": {
      "get": {
        "summary": "List Pages",
        "description": "アーカイブ内のページを viewer と同じ並びで返す",
        "operationId": "list_pages_api_pages_get",
        "parameters": [
          {
            "name": "archive",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Archive"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/thumb": {
      "get": {
        "summary": "Thumbnail",
        "description": "ページのサムネイルを返す",
        "operationId": "thumbnail_api_thumb_get",
        "parameters": [
          {
            "name": "archive",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Archive"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 240,
              "title": "Width"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse": {
      "get": {
        "summary": "Browse",
        "description": "許可された場所の中を辿る。\n\nブラウザはドロップされたファイルの実パスを取得できないため、\nサーバー側で辿って選んでもらう。Tauri ではネイティブのドロップも\n使えるが、同じ画面で両方使えるようにする。",
        "operationId": "browse_api_browse_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Path"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/entries": {
      "get": {
        "summary": "List Entries",
        "description": "タイトルと著者の辞書。query を与えると絞り込む",
        "operationId": "list_entries_api_library_entries_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryEntries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save Entry",
        "description": "辞書に記録する。同じタイトルがあれば上書きする",
        "operationId": "save_entry_api_library_entries_post",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryEntry"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Entry",
        "description": "辞書から取り除く",
        "operationId": "delete_entry_api_library_entries_delete",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Title"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Entry Api Library Entries Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/suggest": {
      "post": {
        "summary": "Suggest",
        "description": "外部サービスから著者名を補完する。\n\nネットワークに出るため失敗しうる。見つからない場合と区別せず、\n空の結果として返して画面を止めない。",
        "operationId": "suggest_api_library_suggest_post",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Suggestion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image": {
      "get": {
        "summary": "Image",
        "description": "ページを原寸で返す。拡大表示に使う",
        "operationId": "image_api_image_get",
        "parameters": [
          {
            "name": "archive",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Archive"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "使い捨てトークン",
              "default": "",
              "title": "Token"
            },
            "description": "使い捨てトークン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrowseResult": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/BrowseEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "path",
          "entries"
        ],
        "title": "BrowseResult",
        "description": "辿っている場所と、その中身"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryEntries": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LibraryEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "LibraryEntries",
        "description": "辞書の中身"
      },
...
      "PageList": {
        "properties": {
          "archive": {
            "type": "string",
            "title": "Archive"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageView"
            },
            "type": "array",
            "title": "Pages"
          },
          "thumbnail_widths": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Thumbnail Widths"
          }
        },
        "type": "object",
        "required": [
          "archive",
          "pages",
          "thumbnail_widths"
        ],
        "title": "PageList",
        "description": "アーカイブ内のページ一覧"
      },
...
    }
  }
}